[SPARK-47454][PYTHON][CONNECT][TESTS] Split `pyspark.sql.tests.test_dataframe`
### What changes were proposed in this pull request?
Split `pyspark.sql.tests.test_dataframe`

### Why are the changes needed?
for testing parallelism

### Does this PR introduce _any_ user-facing change?
no, test only

### How was this patch tested?
updated ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #45580 from zhengruifeng/break_test_df.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
