+++ python/pyspark/sql/tests/test_dataframe.py
@@ -26,8 +26,8 @@ from typing import cast
 import io
 from contextlib import redirect_stdout
 
-from pyspark.sql import SparkSession, Row, functions
-from pyspark.sql.functions import col, lit, count, sum, mean, struct, spark_partition_id
+from pyspark.sql import Row, functions
+from pyspark.sql.functions import col, lit, count, struct, spark_partition_id
 from pyspark.sql.types import (
     StringType,
     IntegerType,
@@ -46,13 +46,11 @@ from pyspark.storagelevel import StorageLevel
 from pyspark.errors import (
     AnalysisException,
     IllegalArgumentException,
-    PySparkAssertionError,
     PySparkTypeError,
     PySparkValueError,
 )
 from pyspark.testing.sqlutils import (
     ReusedSQLTestCase,
-    SQLTestUtils,
     have_pyarrow,
     have_pandas,
     pandas_requirement_message,
@@ -147,13 +145,6 @@ class DataFrameTestsMixin:
         self.assertRaises(AnalysisException, lambda: df.select(df[0]).first())
         self.assertRaises(AnalysisException, lambda: df.select(df.c).first())
         self.assertRaises(AnalysisException, lambda: df.select(df["c"]).first())
-
-    def test_freqItems(self):
-        vals = [Row(a=1, b=-2.0) if i % 2 == 0 else Row(a=i, b=i * 1.0) for i in range(100)]
-        df = self.spark.createDataFrame(vals)
-        items = df.stat.freqItems(("a", "b"), 0.4).collect()[0]
-        self.assertTrue(1 in items[0])
-        self.assertTrue(-2.0 in items[1])
 
     def test_help_command(self):
         # Regression test for SPARK-5464
@@ -315,178 +306,6 @@ class DataFrameTestsMixin:
             df.drop("unknown.unknown").columns, ["id", "first.name", "city.name", "state"]
         )
 
-    def test_dropna(self):
-        schema = StructType(
-            [
-                StructField("name", StringType(), True),
-                StructField("age", IntegerType(), True),
-                StructField("height", DoubleType(), True),
-            ]
-        )
-
-        # shouldn't drop a non-null row
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", 50, 80.1)], schema).dropna().count(), 1
-        )
-
-        # dropping rows with a single null value
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", None, 80.1)], schema).dropna().count(), 0
-        )
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", None, 80.1)], schema).dropna(how="any").count(), 0
-        )
-
-        # if how = 'all', only drop rows if all values are null
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", None, 80.1)], schema).dropna(how="all").count(), 1
-        )
-        self.assertEqual(
-            self.spark.createDataFrame([(None, None, None)], schema).dropna(how="all").count(), 0
-        )
-
-        # how and subset
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", 50, None)], schema)
-            .dropna(how="any", subset=["name", "age"])
-            .count(),
-            1,
-        )
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", None, None)], schema)
-            .dropna(how="any", subset=["name", "age"])
-            .count(),
-            0,
-        )
-
-        # threshold
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", None, 80.1)], schema).dropna(thresh=2).count(), 1
-        )
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", None, None)], schema).dropna(thresh=2).count(), 0
-        )
-
-        # threshold and subset
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", 50, None)], schema)
-            .dropna(thresh=2, subset=["name", "age"])
-            .count(),
-            1,
-        )
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", None, 180.9)], schema)
-            .dropna(thresh=2, subset=["name", "age"])
-            .count(),
-            0,
-        )
-
-        # thresh should take precedence over how
-        self.assertEqual(
-            self.spark.createDataFrame([("Alice", 50, None)], schema)
-            .dropna(how="any", thresh=2, subset=["name", "age"])
-            .count(),
-            1,
-        )
-
-        with self.assertRaises(PySparkTypeError) as pe:
-            self.spark.createDataFrame([("Alice", 50, None)], schema).dropna(subset=10)
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="NOT_LIST_OR_STR_OR_TUPLE",
-            message_parameters={"arg_name": "subset", "arg_type": "int"},
-        )
-
-    def test_fillna(self):
-        schema = StructType(
-            [
-                StructField("name", StringType(), True),
-                StructField("age", IntegerType(), True),
-                StructField("height", DoubleType(), True),
-                StructField("spy", BooleanType(), True),
-            ]
-        )
-
-        # fillna shouldn't change non-null values
-        row = self.spark.createDataFrame([("Alice", 10, 80.1, True)], schema).fillna(50).first()
-        self.assertEqual(row.age, 10)
-
-        # fillna with int
-        row = self.spark.createDataFrame([("Alice", None, None, None)], schema).fillna(50).first()
-        self.assertEqual(row.age, 50)
-        self.assertEqual(row.height, 50.0)
-
-        # fillna with double
-        row = self.spark.createDataFrame([("Alice", None, None, None)], schema).fillna(50.1).first()
-        self.assertEqual(row.age, 50)
-        self.assertEqual(row.height, 50.1)
-
-        # fillna with bool
-        row = self.spark.createDataFrame([("Alice", None, None, None)], schema).fillna(True).first()
-        self.assertEqual(row.age, None)
-        self.assertEqual(row.spy, True)
-
-        # fillna with string
-        row = self.spark.createDataFrame([(None, None, None, None)], schema).fillna("hello").first()
-        self.assertEqual(row.name, "hello")
-        self.assertEqual(row.age, None)
-
-        # fillna with subset specified for numeric cols
-        row = (
-            self.spark.createDataFrame([(None, None, None, None)], schema)
-            .fillna(50, subset=["name", "age"])
-            .first()
-        )
-        self.assertEqual(row.name, None)
-        self.assertEqual(row.age, 50)
-        self.assertEqual(row.height, None)
-        self.assertEqual(row.spy, None)
-
-        # fillna with subset specified for string cols
-        row = (
-            self.spark.createDataFrame([(None, None, None, None)], schema)
-            .fillna("haha", subset=["name", "age"])
-            .first()
-        )
-        self.assertEqual(row.name, "haha")
-        self.assertEqual(row.age, None)
-        self.assertEqual(row.height, None)
-        self.assertEqual(row.spy, None)
-
-        # fillna with subset specified for bool cols
-        row = (
-            self.spark.createDataFrame([(None, None, None, None)], schema)
-            .fillna(True, subset=["name", "spy"])
-            .first()
-        )
-        self.assertEqual(row.name, None)
-        self.assertEqual(row.age, None)
-        self.assertEqual(row.height, None)
-        self.assertEqual(row.spy, True)
-
-        # fillna with dictionary for boolean types
-        row = self.spark.createDataFrame([Row(a=None), Row(a=True)]).fillna({"a": True}).first()
-        self.assertEqual(row.a, True)
-
-        with self.assertRaises(PySparkTypeError) as pe:
-            self.spark.createDataFrame([Row(a=None), Row(a=True)]).fillna(["a", True])
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="NOT_BOOL_OR_DICT_OR_FLOAT_OR_INT_OR_STR",
-            message_parameters={"arg_name": "value", "arg_type": "list"},
-        )
-
-        with self.assertRaises(PySparkTypeError) as pe:
-            self.spark.createDataFrame([Row(a=None), Row(a=True)]).fillna(50, subset=10)
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="NOT_LIST_OR_TUPLE",
-            message_parameters={"arg_name": "subset", "arg_type": "int"},
-        )
-
     def test_repartitionByRange_dataframe(self):
         schema = StructType(
             [
@@ -528,205 +347,6 @@ class DataFrameTestsMixin:
             exception=pe.exception,
             error_class="NOT_COLUMN_OR_INT_OR_STR",
             message_parameters={"arg_name": "numPartitions", "arg_type": "list"},
-        )
-
-    def test_replace(self):
-        schema = StructType(
-            [
-                StructField("name", StringType(), True),
-                StructField("age", IntegerType(), True),
-                StructField("height", DoubleType(), True),
-            ]
-        )
-
-        # replace with int
-        row = self.spark.createDataFrame([("Alice", 10, 10.0)], schema).replace(10, 20).first()
-        self.assertEqual(row.age, 20)
-        self.assertEqual(row.height, 20.0)
-
-        # replace with double
-        row = self.spark.createDataFrame([("Alice", 80, 80.0)], schema).replace(80.0, 82.1).first()
-        self.assertEqual(row.age, 82)
-        self.assertEqual(row.height, 82.1)
-
-        # replace with string
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema)
-            .replace("Alice", "Ann")
-            .first()
-        )
-        self.assertEqual(row.name, "Ann")
-        self.assertEqual(row.age, 10)
-
-        # replace with subset specified by a string of a column name w/ actual change
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema)
-            .replace(10, 20, subset="age")
-            .first()
-        )
-        self.assertEqual(row.age, 20)
-
-        # replace with subset specified by a string of a column name w/o actual change
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema)
-            .replace(10, 20, subset="height")
-            .first()
-        )
-        self.assertEqual(row.age, 10)
-
-        # replace with subset specified with one column replaced, another column not in subset
-        # stays unchanged.
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 10.0)], schema)
-            .replace(10, 20, subset=["name", "age"])
-            .first()
-        )
-        self.assertEqual(row.name, "Alice")
-        self.assertEqual(row.age, 20)
-        self.assertEqual(row.height, 10.0)
-
-        # replace with subset specified but no column will be replaced
-        row = (
-            self.spark.createDataFrame([("Alice", 10, None)], schema)
-            .replace(10, 20, subset=["name", "height"])
-            .first()
-        )
-        self.assertEqual(row.name, "Alice")
-        self.assertEqual(row.age, 10)
-        self.assertEqual(row.height, None)
-
-        # replace with lists
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema)
-            .replace(["Alice"], ["Ann"])
-            .first()
-        )
-        self.assertTupleEqual(row, ("Ann", 10, 80.1))
-
-        # replace with dict
-        row = self.spark.createDataFrame([("Alice", 10, 80.1)], schema).replace({10: 11}).first()
-        self.assertTupleEqual(row, ("Alice", 11, 80.1))
-
-        # test backward compatibility with dummy value
-        dummy_value = 1
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema)
-            .replace({"Alice": "Bob"}, dummy_value)
-            .first()
-        )
-        self.assertTupleEqual(row, ("Bob", 10, 80.1))
-
-        # test dict with mixed numerics
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema)
-            .replace({10: -10, 80.1: 90.5})
-            .first()
-        )
-        self.assertTupleEqual(row, ("Alice", -10, 90.5))
-
-        # replace with tuples
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema)
-            .replace(("Alice",), ("Bob",))
-            .first()
-        )
-        self.assertTupleEqual(row, ("Bob", 10, 80.1))
-
-        # replace multiple columns
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.0)], schema)
-            .replace((10, 80.0), (20, 90))
-            .first()
-        )
-        self.assertTupleEqual(row, ("Alice", 20, 90.0))
-
-        # test for mixed numerics
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.0)], schema)
-            .replace((10, 80), (20, 90.5))
-            .first()
-        )
-        self.assertTupleEqual(row, ("Alice", 20, 90.5))
-
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.0)], schema)
-            .replace({10: 20, 80: 90.5})
-            .first()
-        )
-        self.assertTupleEqual(row, ("Alice", 20, 90.5))
-
-        # replace with boolean
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.0)], schema)
-            .selectExpr("name = 'Bob'", "age <= 15")
-            .replace(False, True)
-            .first()
-        )
-        self.assertTupleEqual(row, (True, True))
-
-        # replace string with None and then drop None rows
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.0)], schema)
-            .replace("Alice", None)
-            .dropna()
-        )
-        self.assertEqual(row.count(), 0)
-
-        # replace with number and None
-        row = (
-            self.spark.createDataFrame([("Alice", 10, 80.0)], schema)
-            .replace([10, 80], [20, None])
-            .first()
-        )
-        self.assertTupleEqual(row, ("Alice", 20, None))
-
-        # should fail if subset is not list, tuple or None
-        with self.assertRaises(TypeError):
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema).replace(
-                {10: 11}, subset=1
-            ).first()
-
-        # should fail if to_replace and value have different length
-        with self.assertRaises(ValueError):
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema).replace(
-                ["Alice", "Bob"], ["Eve"]
-            ).first()
-
-        # should fail if when received unexpected type
-        with self.assertRaises(TypeError):
-            from datetime import datetime
-
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema).replace(
-                datetime.now(), datetime.now()
-            ).first()
-
-        # should fail if provided mixed type replacements
-        with self.assertRaises(ValueError):
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema).replace(
-                ["Alice", 10], ["Eve", 20]
-            ).first()
-
-        with self.assertRaises(ValueError):
-            self.spark.createDataFrame([("Alice", 10, 80.1)], schema).replace(
-                {"Alice": "Bob", 10: 20}
-            ).first()
-
-        with self.assertRaises(PySparkTypeError) as pe:
-            self.spark.createDataFrame([("Alice", 10, 80.0)], schema).replace(["Alice", "Bob"])
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="ARGUMENT_REQUIRED",
-            message_parameters={"arg_name": "value", "condition": "`to_replace` is dict"},
-        )
-
-        with self.assertRaises(PySparkTypeError) as pe:
-            self.spark.createDataFrame([("Alice", 10, 80.0)], schema).replace(lambda x: x + 1, 10)
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="NOT_BOOL_OR_DICT_OR_FLOAT_OR_INT_OR_LIST_OR_STR_OR_TUPLE",
-            message_parameters={"arg_name": "to_replace", "arg_type": "function"},
         )
 
     def test_with_column_with_existing_name(self):
@@ -833,408 +453,6 @@ class DataFrameTestsMixin:
             self.assertGreaterEqual(explain_output.count("1.2345"), 1)
             self.assertGreaterEqual(explain_output.count("what"), 1)
             self.assertGreaterEqual(explain_output.count("itworks"), 1)
-
-    def test_unpivot(self):
-        # SPARK-39877: test the DataFrame.unpivot method
-        df = self.spark.createDataFrame(
-            [
-                (1, 10, 1.0, "one"),
-                (2, 20, 2.0, "two"),
-                (3, 30, 3.0, "three"),
-            ],
-            ["id", "int", "double", "str"],
-        )
-
-        with self.subTest(desc="with none identifier"):
-            with self.assertRaisesRegex(AssertionError, "ids must not be None"):
-                df.unpivot(None, ["int", "double"], "var", "val")
-
-        with self.subTest(desc="with no identifier"):
-            for id in [[], ()]:
-                with self.subTest(ids=id):
-                    actual = df.unpivot(id, ["int", "double"], "var", "val")
-                    self.assertEqual(actual.schema.simpleString(), "struct<var:string,val:double>")
-                    self.assertEqual(
-                        actual.collect(),
-                        [
-                            Row(var="int", value=10.0),
-                            Row(var="double", value=1.0),
-                            Row(var="int", value=20.0),
-                            Row(var="double", value=2.0),
-                            Row(var="int", value=30.0),
-                            Row(var="double", value=3.0),
-                        ],
-                    )
-
-        with self.subTest(desc="with single identifier column"):
-            for id in ["id", ["id"], ("id",)]:
-                with self.subTest(ids=id):
-                    actual = df.unpivot(id, ["int", "double"], "var", "val")
-                    self.assertEqual(
-                        actual.schema.simpleString(),
-                        "struct<id:bigint,var:string,val:double>",
-                    )
-                    self.assertEqual(
-                        actual.collect(),
-                        [
-                            Row(id=1, var="int", value=10.0),
-                            Row(id=1, var="double", value=1.0),
-                            Row(id=2, var="int", value=20.0),
-                            Row(id=2, var="double", value=2.0),
-                            Row(id=3, var="int", value=30.0),
-                            Row(id=3, var="double", value=3.0),
-                        ],
-                    )
-
-        with self.subTest(desc="with multiple identifier columns"):
-            for ids in [["id", "double"], ("id", "double")]:
-                with self.subTest(ids=ids):
-                    actual = df.unpivot(ids, ["int", "double"], "var", "val")
-                    self.assertEqual(
-                        actual.schema.simpleString(),
-                        "struct<id:bigint,double:double,var:string,val:double>",
-                    )
-                    self.assertEqual(
-                        actual.collect(),
-                        [
-                            Row(id=1, double=1.0, var="int", value=10.0),
-                            Row(id=1, double=1.0, var="double", value=1.0),
-                            Row(id=2, double=2.0, var="int", value=20.0),
-                            Row(id=2, double=2.0, var="double", value=2.0),
-                            Row(id=3, double=3.0, var="int", value=30.0),
-                            Row(id=3, double=3.0, var="double", value=3.0),
-                        ],
-                    )
-
-        with self.subTest(desc="with no identifier columns but none value columns"):
-            # select only columns that have common data type (double)
-            actual = df.select("id", "int", "double").unpivot([], None, "var", "val")
-            self.assertEqual(actual.schema.simpleString(), "struct<var:string,val:double>")
-            self.assertEqual(
-                actual.collect(),
-                [
-                    Row(var="id", value=1.0),
-                    Row(var="int", value=10.0),
-                    Row(var="double", value=1.0),
-                    Row(var="id", value=2.0),
-                    Row(var="int", value=20.0),
-                    Row(var="double", value=2.0),
-                    Row(var="id", value=3.0),
-                    Row(var="int", value=30.0),
-                    Row(var="double", value=3.0),
-                ],
-            )
-
-        with self.subTest(desc="with single identifier columns but none value columns"):
-            for ids in ["id", ["id"], ("id",)]:
-                with self.subTest(ids=ids):
-                    # select only columns that have common data type (double)
-                    actual = df.select("id", "int", "double").unpivot(ids, None, "var", "val")
-                    self.assertEqual(
-                        actual.schema.simpleString(), "struct<id:bigint,var:string,val:double>"
-                    )
-                    self.assertEqual(
-                        actual.collect(),
-                        [
-                            Row(id=1, var="int", value=10.0),
-                            Row(id=1, var="double", value=1.0),
-                            Row(id=2, var="int", value=20.0),
-                            Row(id=2, var="double", value=2.0),
-                            Row(id=3, var="int", value=30.0),
-                            Row(id=3, var="double", value=3.0),
-                        ],
-                    )
-
-        with self.subTest(desc="with multiple identifier columns but none given value columns"):
-            for ids in [["id", "str"], ("id", "str")]:
-                with self.subTest(ids=ids):
-                    actual = df.unpivot(ids, None, "var", "val")
-                    self.assertEqual(
-                        actual.schema.simpleString(),
-                        "struct<id:bigint,str:string,var:string,val:double>",
-                    )
-                    self.assertEqual(
-                        actual.collect(),
-                        [
-                            Row(id=1, str="one", var="int", val=10.0),
-                            Row(id=1, str="one", var="double", val=1.0),
-                            Row(id=2, str="two", var="int", val=20.0),
-                            Row(id=2, str="two", var="double", val=2.0),
-                            Row(id=3, str="three", var="int", val=30.0),
-                            Row(id=3, str="three", var="double", val=3.0),
-                        ],
-                    )
-
-        with self.subTest(desc="with single value column"):
-            for values in ["int", ["int"], ("int",)]:
-                with self.subTest(values=values):
-                    actual = df.unpivot("id", values, "var", "val")
-                    self.assertEqual(
-                        actual.schema.simpleString(), "struct<id:bigint,var:string,val:bigint>"
-                    )
-                    self.assertEqual(
-                        actual.collect(),
-                        [
-                            Row(id=1, var="int", val=10),
-                            Row(id=2, var="int", val=20),
-                            Row(id=3, var="int", val=30),
-                        ],
-                    )
-
-        with self.subTest(desc="with multiple value columns"):
-            for values in [["int", "double"], ("int", "double")]:
-                with self.subTest(values=values):
-                    actual = df.unpivot("id", values, "var", "val")
-                    self.assertEqual(
-                        actual.schema.simpleString(), "struct<id:bigint,var:string,val:double>"
-                    )
-                    self.assertEqual(
-                        actual.collect(),
-                        [
-                            Row(id=1, var="int", val=10.0),
-                            Row(id=1, var="double", val=1.0),
-                            Row(id=2, var="int", val=20.0),
-                            Row(id=2, var="double", val=2.0),
-                            Row(id=3, var="int", val=30.0),
-                            Row(id=3, var="double", val=3.0),
-                        ],
-                    )
-
-        with self.subTest(desc="with columns"):
-            for id in [df.id, [df.id], (df.id,)]:
-                for values in [[df.int, df.double], (df.int, df.double)]:
-                    with self.subTest(ids=id, values=values):
-                        self.assertEqual(
-                            df.unpivot(id, values, "var", "val").collect(),
-                            df.unpivot("id", ["int", "double"], "var", "val").collect(),
-                        )
-
-        with self.subTest(desc="with column names and columns"):
-            for ids in [[df.id, "str"], (df.id, "str")]:
-                for values in [[df.int, "double"], (df.int, "double")]:
-                    with self.subTest(ids=ids, values=values):
-                        self.assertEqual(
-                            df.unpivot(ids, values, "var", "val").collect(),
-                            df.unpivot(["id", "str"], ["int", "double"], "var", "val").collect(),
-                        )
-
-        with self.subTest(desc="melt alias"):
-            self.assertEqual(
-                df.unpivot("id", ["int", "double"], "var", "val").collect(),
-                df.melt("id", ["int", "double"], "var", "val").collect(),
-            )
-
-    def test_unpivot_negative(self):
-        # SPARK-39877: test the DataFrame.unpivot method
-        df = self.spark.createDataFrame(
-            [
-                (1, 10, 1.0, "one"),
-                (2, 20, 2.0, "two"),
-                (3, 30, 3.0, "three"),
-            ],
-            ["id", "int", "double", "str"],
-        )
-
-        with self.subTest(desc="with no value columns"):
-            for values in [[], ()]:
-                with self.subTest(values=values):
-                    with self.assertRaisesRegex(
-                        AnalysisException,
-                        r"\[UNPIVOT_REQUIRES_VALUE_COLUMNS] At least one value column "
-                        r"needs to be specified for UNPIVOT, all columns specified as ids.*",
-                    ):
-                        df.unpivot("id", values, "var", "val").collect()
-
-        with self.subTest(desc="with value columns without common data type"):
-            with self.assertRaisesRegex(
-                AnalysisException,
-                r"\[UNPIVOT_VALUE_DATA_TYPE_MISMATCH\] Unpivot value columns must share "
-                r"a least common type, some types do not: .*",
-            ):
-                df.unpivot("id", ["int", "str"], "var", "val").collect()
-
-    def test_melt_groupby(self):
-        df = self.spark.createDataFrame(
-            [(1, 2, 3, 4, 5, 6)],
-            ["f1", "f2", "label", "pred", "model_version", "ts"],
-        )
-        self.assertEqual(
-            df.melt(
-                "model_version",
-                ["label", "f2"],
-                "f1",
-                "f2",
-            )
-            .groupby("f1")
-            .count()
-            .count(),
-            2,
-        )
-
-    def test_observe(self):
-        # SPARK-36263: tests the DataFrame.observe(Observation, *Column) method
-        from pyspark.sql import Observation
-
-        df = self.spark.createDataFrame(
-            [
-                (1, 1.0, "one"),
-                (2, 2.0, "two"),
-                (3, 3.0, "three"),
-            ],
-            ["id", "val", "label"],
-        )
-
-        unnamed_observation = Observation()
-        named_observation = Observation("metric")
-
-        with self.assertRaises(PySparkAssertionError) as pe:
-            unnamed_observation.get()
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="NO_OBSERVE_BEFORE_GET",
-            message_parameters={},
-        )
-
-        observed = (
-            df.orderBy("id")
-            .observe(
-                named_observation,
-                count(lit(1)).alias("cnt"),
-                sum(col("id")).alias("sum"),
-                mean(col("val")).alias("mean"),
-            )
-            .observe(unnamed_observation, count(lit(1)).alias("rows"))
-        )
-
-        # test that observe works transparently
-        actual = observed.collect()
-        self.assertEqual(
-            [
-                {"id": 1, "val": 1.0, "label": "one"},
-                {"id": 2, "val": 2.0, "label": "two"},
-                {"id": 3, "val": 3.0, "label": "three"},
-            ],
-            [row.asDict() for row in actual],
-        )
-
-        # test that we retrieve the metrics
-        self.assertEqual(named_observation.get, dict(cnt=3, sum=6, mean=2.0))
-        self.assertEqual(unnamed_observation.get, dict(rows=3))
-
-        with self.assertRaises(PySparkAssertionError) as pe:
-            df.observe(named_observation, count(lit(1)).alias("count"))
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="REUSE_OBSERVATION",
-            message_parameters={},
-        )
-
-        # observation requires name (if given) to be non empty string
-        with self.assertRaisesRegex(TypeError, "`name` should be a str, got int"):
-            Observation(123)
-        with self.assertRaisesRegex(ValueError, "`name` must be a non-empty string, got ''."):
-            Observation("")
-
-        # dataframe.observe requires at least one expr
-        with self.assertRaises(PySparkValueError) as pe:
-            df.observe(Observation())
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="CANNOT_BE_EMPTY",
-            message_parameters={"item": "exprs"},
-        )
-
-        # dataframe.observe requires non-None Columns
-        for args in [(None,), ("id",), (lit(1), None), (lit(1), "id")]:
-            with self.subTest(args=args):
-                with self.assertRaises(PySparkTypeError) as pe:
-                    df.observe(Observation(), *args)
-
-                self.check_error(
-                    exception=pe.exception,
-                    error_class="NOT_LIST_OF_COLUMN",
-                    message_parameters={"arg_name": "exprs"},
-                )
-
-    def test_observe_str(self):
-        # SPARK-38760: tests the DataFrame.observe(str, *Column) method
-        from pyspark.sql.streaming import StreamingQueryListener
-
-        observed_metrics = None
-
-        class TestListener(StreamingQueryListener):
-            def onQueryStarted(self, event):
-                pass
-
-            def onQueryProgress(self, event):
-                nonlocal observed_metrics
-                observed_metrics = event.progress.observedMetrics
-
-            def onQueryIdle(self, event):
-                pass
-
-            def onQueryTerminated(self, event):
-                pass
-
-        self.spark.streams.addListener(TestListener())
-
-        df = self.spark.readStream.format("rate").option("rowsPerSecond", 10).load()
-        df = df.observe("metric", count(lit(1)).alias("cnt"), sum(col("value")).alias("sum"))
-        q = df.writeStream.format("noop").queryName("test").start()
-        self.assertTrue(q.isActive)
-        time.sleep(10)
-        q.stop()
-
-        self.assertTrue(isinstance(observed_metrics, dict))
-        self.assertTrue("metric" in observed_metrics)
-        row = observed_metrics["metric"]
-        self.assertTrue(isinstance(row, Row))
-        self.assertTrue(hasattr(row, "cnt"))
-        self.assertTrue(hasattr(row, "sum"))
-        self.assertGreaterEqual(row.cnt, 0)
-        self.assertGreaterEqual(row.sum, 0)
-
-    def test_observe_with_same_name_on_different_dataframe(self):
-        # SPARK-45656: named observations with the same name on different datasets
-        from pyspark.sql import Observation
-
-        observation1 = Observation("named")
-        df1 = self.spark.range(50)
-        observed_df1 = df1.observe(observation1, count(lit(1)).alias("cnt"))
-
-        observation2 = Observation("named")
-        df2 = self.spark.range(100)
-        observed_df2 = df2.observe(observation2, count(lit(1)).alias("cnt"))
-
-        observed_df1.collect()
-        observed_df2.collect()
-
-        self.assertEqual(observation1.get, dict(cnt=50))
-        self.assertEqual(observation2.get, dict(cnt=100))
-
-    def test_observe_on_commands(self):
-        from pyspark.sql import Observation
-
-        df = self.spark.range(50)
-
-        test_table = "test_table"
-
-        # DataFrameWriter
-        with self.table(test_table):
-            for command, action in [
-                ("collect", lambda df: df.collect()),
-                ("show", lambda df: df.show(50)),
-                ("save", lambda df: df.write.format("noop").mode("overwrite").save()),
-                ("create", lambda df: df.writeTo(test_table).using("parquet").create()),
-            ]:
-                with self.subTest(command=command):
-                    observation = Observation()
-                    observed_df = df.observe(observation, count(lit(1)).alias("cnt"))
-                    action(observed_df)
-                    self.assertEqual(observation.get, dict(cnt=50))
 
     def test_sample(self):
         with self.assertRaises(PySparkTypeError) as pe:
@@ -2068,80 +1286,6 @@ class DataFrameTestsMixin:
             error_class="INVALID_TYPE",
             message_parameters={"arg_name": "data", "arg_type": "DataFrame"},
         )
-
-
-class QueryExecutionListenerTests(unittest.TestCase, SQLTestUtils):
-    # These tests are separate because it uses 'spark.sql.queryExecutionListeners' which is
-    # static and immutable. This can't be set or unset, for example, via `spark.conf`.
-
-    @classmethod
-    def setUpClass(cls):
-        import glob
-        from pyspark.find_spark_home import _find_spark_home
-
-        SPARK_HOME = _find_spark_home()
-        filename_pattern = (
-            "sql/core/target/scala-*/test-classes/org/apache/spark/sql/"
-            "TestQueryExecutionListener.class"
-        )
-        cls.has_listener = bool(glob.glob(os.path.join(SPARK_HOME, filename_pattern)))
-
-        if cls.has_listener:
-            # Note that 'spark.sql.queryExecutionListeners' is a static immutable configuration.
-            cls.spark = (
-                SparkSession.builder.master("local[4]")
-                .appName(cls.__name__)
-                .config(
-                    "spark.sql.queryExecutionListeners",
-                    "org.apache.spark.sql.TestQueryExecutionListener",
-                )
-                .getOrCreate()
-            )
-
-    def setUp(self):
-        if not self.has_listener:
-            raise self.skipTest(
-                "'org.apache.spark.sql.TestQueryExecutionListener' is not "
-                "available. Will skip the related tests."
-            )
-
-    @classmethod
-    def tearDownClass(cls):
-        if hasattr(cls, "spark"):
-            cls.spark.stop()
-
-    def tearDown(self):
-        self.spark._jvm.OnSuccessCall.clear()
-
-    def test_query_execution_listener_on_collect(self):
-        self.assertFalse(
-            self.spark._jvm.OnSuccessCall.isCalled(),
-            "The callback from the query execution listener should not be called before 'collect'",
-        )
-        self.spark.sql("SELECT * FROM range(1)").collect()
-        self.spark.sparkContext._jsc.sc().listenerBus().waitUntilEmpty(10000)
-        self.assertTrue(
-            self.spark._jvm.OnSuccessCall.isCalled(),
-            "The callback from the query execution listener should be called after 'collect'",
-        )
-
-    @unittest.skipIf(
-        not have_pandas or not have_pyarrow,
-        cast(str, pandas_requirement_message or pyarrow_requirement_message),
-    )
-    def test_query_execution_listener_on_collect_with_arrow(self):
-        with self.sql_conf({"spark.sql.execution.arrow.pyspark.enabled": True}):
-            self.assertFalse(
-                self.spark._jvm.OnSuccessCall.isCalled(),
-                "The callback from the query execution listener should not be "
-                "called before 'toPandas'",
-            )
-            self.spark.sql("SELECT * FROM range(1)").toPandas()
-            self.spark.sparkContext._jsc.sc().listenerBus().waitUntilEmpty(10000)
-            self.assertTrue(
-                self.spark._jvm.OnSuccessCall.isCalled(),
-                "The callback from the query execution listener should be called after 'toPandas'",
-            )
 
 
 class DataFrameTests(DataFrameTestsMixin, ReusedSQLTestCase):
